Move theme property translations to own file
